+++ src/stdio/remove.c
@@ -1,16 +1,16 @@
 #include <stdlib.h>
 
-#include <los/filesystem.h>
 #include <los/error.h>
+#include <los/filesystem.h>
 
 int remove(const char* filename) {
-    int64_t status = remove_file(filename);
-    if(status == EISDIR) {
-        if(remove_directory(filename) < 0)
+    isize status = remove_file(filename);
+    if (status == EISDIR) {
+        if (remove_directory(filename) < 0)
             return -1;
-        
+
         return 0;
-    } else if(status < 0)
+    } else if (status < 0)
         return -1;
 
     return 0;
